+++ appnav/src/main/kotlin/io/element/android/appnav/session/MatrixSessionCache.kt
@@ -6,7 +6,7 @@
  * Please see LICENSE files in the repository root for full details.
  */
 
-package io.element.android.appnav.di
+package io.element.android.appnav.session
 
 import androidx.annotation.VisibleForTesting
 import com.bumble.appyx.core.state.MutableSavedStateMap
